ddl sql for different storage.

+++ packages/watchmen-storage-oracle/meta-scripts/16.5.19/00002-change_data_record_history.ddl.sql
@@ -1,2 +1,2 @@
 ALTER TABLE change_data_record_history
-    ADD status BIGINT;
+    ADD status NUMBER(1);
