+++ packages/watchmen-storage-oracle/meta-scripts/16.5.19/00003-change_data_json.ddl.sql
@@ -1,2 +1,2 @@
 ALTER TABLE change_data_json
-    ADD status BIGINT;
+    ADD status NUMBER(1);
